Added background

+++ src/com/psywerx/dh/Texture.java
@@ -20,8 +20,8 @@ public class Texture {
 
     public float[] getTextureUV() {
         float[] charWidth = { size[0]/64.0f, size[1]/64.0f };
-        float u = (sprite[0] / size[0]);
-        float v = (sprite[1] / size[1]);
+        float u = (sprite[0] / (float)size[0]);
+        float v = (sprite[1] / (float)size[1]);
         return new float[] { (u + 1) * charWidth[0], (v + 1) * charWidth[1],
                 u * charWidth[0], (v + 1) * charWidth[1],
                 (u + 1f) * charWidth[0], v * charWidth[1], u * charWidth[0],
